+++ backend/src/main/java/com/wellbeignatwork/backend/Service/Forum/ReactionService.java
@@ -3,6 +3,6 @@ package com.wellbeignatwork.backend.service.Forum;
 import com.wellbeignatwork.backend.entity.Forum.Reaction;
 
 public interface ReactionService {
-    public void addReactToPost(Reaction reaction,int idPost,int idUser);
+    public void addReactToPost(Reaction reaction,int idPost,Long idUser);
     public void deleteReaction(int idReaction);
 }
